Avoid duplicating test class
